// src/lib/http/generated/models/search-companies-response-dto-output-items-item.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API de Cuidados com Animais
 * OpenAPI spec version: 1.0
 */

import type { SearchCompaniesResponseDtoOutputItemsItemAddress } from "./search-companies-response-dto-output-items-item-address";
import type { SearchCompaniesResponseDtoOutputItemsItemAverageRating } from "./search-companies-response-dto-output-items-item-average-rating";
import type { SearchCompaniesResponseDtoOutputItemsItemContact } from "./search-companies-response-dto-output-items-item-contact";
import type { SearchCompaniesResponseDtoOutputItemsItemDescription } from "./search-companies-response-dto-output-items-item-description";
import type { SearchCompaniesResponseDtoOutputItemsItemImage } from "./search-companies-response-dto-output-items-item-image";
import type { SearchCompaniesResponseDtoOutputItemsItemRatingCount } from "./search-companies-response-dto-output-items-item-rating-count";

export type SearchCompaniesResponseDtoOutputItemsItem = {
	id: string;
	name: string;
	contact?: SearchCompaniesResponseDtoOutputItemsItemContact;
	averageRating?: SearchCompaniesResponseDtoOutputItemsItemAverageRating;
	ratingCount?: SearchCompaniesResponseDtoOutputItemsItemRatingCount;
	description?: SearchCompaniesResponseDtoOutputItemsItemDescription;
	address: SearchCompaniesResponseDtoOutputItemsItemAddress;
	image?: SearchCompaniesResponseDtoOutputItemsItemImage;
};